backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cffi==1.17.1
contextlib2==21.6.0
cryptography==44.0.3
Django==4.2.21
django-ninja==1.4.1
django-ninja-extra==0.30.0
django-ninja-jwt==5.3.7
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
idna==3.10
iniconfig==2.1.0
injector==0.22.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
pytest==8.0.2
pytest-django==4.8.0
python-dotenv==1.1.0
sqlparse==0.5.3
tomli==2.2.1
typing-inspection==0.4.0
typing_extensions==4.13.2
